+++ src/main/java/br/com/invistaix/InvistaIX/repository/ProprietarioRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import br.com.invistaix.InvistaIX.model.ProprietarioModel;
 
 @Repository
-public interface ProprietarioRepository  extends JpaRepository<ProprietarioModel, Integer> {
+public interface ProprietarioRepository  extends JpaRepository<ProprietarioModel, Long> {
 	@Query("SELECT p FROM ProprietarioModel p WHERE p.cnpjCpf = :cnpjCpf")
     Optional<ProprietarioModel> findByCnpjCpf(@Param("cnpjCpf") String cnpjCpf);
 }
